src/requirements.txt
Flask==2.2.2
SQLAlchemy==1.4.40
gunicorn==20.1.0

PyMySQL==1.0.2
coloredlogs==15.0.1
python-dotenv==0.20.0
numpy==1.23.3
torch==1.12.1
tqdm==4.64.1
more-itertools==8.14.0
transformers>=4.19.0
ffmpeg-python==0.2.0
